+++ src/sizes.rs
@@ -51,26 +51,26 @@ pub const MAX_CHANNEL_VALUE: u8 = 255;
 
 // convert a block or chunk coord to its containing chunk or region coord
 
-fn divide(value: i32, ratio: usize) -> i32 {
-    (value as f64 / ratio as f64).floor() as i32
+fn divide(value: isize, ratio: usize) -> isize {
+    (value as f64 / ratio as f64).floor() as isize
 }
-pub fn block_to_chunk(b: i32) -> i32 {
+pub fn block_to_chunk(b: isize) -> isize {
     divide(b, BLOCKS_IN_CHUNK)
 }
-pub fn block_to_region(b: i32) -> i32 {
+pub fn block_to_region(b: isize) -> isize {
     divide(b, BLOCKS_IN_REGION)
 }
-pub fn chunk_to_region(c: i32) -> i32 {
+pub fn chunk_to_region(c: isize) -> isize {
     divide(c, CHUNKS_IN_REGION)
 }
 
 
 // convert a block or chunk coord to its local coord within a chunk or region
 
-fn signed_modulo(value: i32, ratio: usize) -> usize {
+fn signed_modulo(value: isize, ratio: usize) -> usize {
     (value as f64 % ratio as f64 + ratio as f64) as usize % ratio
 }
-pub fn block_pos_in_chunk(b: i32, c: Option<i32>) -> usize {
+pub fn block_pos_in_chunk(b: isize, c: Option<isize>) -> usize {
     match c {
         Some(c) => match block_to_chunk(b) {
             // If the block doesn't fall within the given chunk, clamp to the edge of the chunk.
@@ -81,10 +81,10 @@ pub fn block_pos_in_chunk(b: i32, c: Option<i32>) -> usize {
         None => signed_modulo(b, BLOCKS_IN_CHUNK),
     }
 }
-// pub fn block_pos_in_region(b: i32) -> usize {
+// pub fn block_pos_in_region(b: isize) -> usize {
 //     signed_modulo(b, BLOCKS_IN_REGION)
 // }
-pub fn chunk_pos_in_region(c: i32, r: Option<i32>) -> usize {
+pub fn chunk_pos_in_region(c: isize, r: Option<isize>) -> usize {
     match r {
         Some(r) => match chunk_to_region(c) {
             // If the chunk doesn't fall within the given region, clamp to the edge of the region.
